fix mac build

+++ src/mac.rs
@@ -8,13 +8,13 @@ use eyre::Context;
 use launchd::Launchd;
 
 use crate::{
-    service_config::ServiceConfig,
+    service_builder::{ServiceBuilder, ServiceLevel},
     service_manager::{Result, ServiceManager},
     service_status::ServiceStatus,
 };
 
 pub struct Manager {
-    config: ServiceConfig,
+    config: ServiceBuilder,
 }
 
 impl Manager {
@@ -32,26 +32,48 @@ impl Manager {
         } else {
             output.stderr
         };
-        let out = String::from_utf8(out_bytes).wrap_err("Error reading output")?;
+        let out = String::from_utf8(out_bytes)
+            .wrap_err("Error reading output")?
+            .trim()
+            .to_lowercase()
+            .to_owned();
         Ok(out)
     }
 
-    fn get_plist_path(&self) -> PathBuf {
-        PathBuf::from("/Library/LaunchDaemons").join(format!("{}.plist", self.config.name))
+    fn user_agent_dir(&self) -> Result<PathBuf> {
+        let user_dirs = directories::UserDirs::new().ok_or("User dirs not found")?;
+        let home_dir = user_dirs.home_dir();
+        Ok(home_dir.join("Library").join("LaunchAgents"))
+    }
+
+    fn get_plist_path(&self) -> Result<PathBuf> {
+        let path = match self.config.service_level {
+            ServiceLevel::System => PathBuf::from("/Library/LaunchDaemons"),
+            ServiceLevel::User => self.user_agent_dir()?,
+        };
+        Ok(path.join(format!("{}.plist", self.config.name)))
     }
 }
 
 impl ServiceManager for Manager {
-    fn new(config: ServiceConfig) -> Result<Self> {
-        Ok(Self { config })
+    fn builder(name: impl Into<String>) -> ServiceBuilder {
+        ServiceBuilder::new(name)
+    }
+
+    fn new(name: impl Into<String>) -> Result<Self> {
+        ServiceBuilder::new(name).build()
+    }
+
+    fn from_builder(builder: ServiceBuilder) -> Result<Self> {
+        Ok(Self { config: builder })
     }
 
     fn install(&self) -> Result<()> {
         let file = Launchd::new(&self.config.name, &self.config.program)
             .wrap_err("Error creating config")?
-            .with_program_arguments(self.config.args_iter().map(|a| a.to_owned()).collect());
+            .with_program_arguments(self.config.full_args_iter().map(|a| a.to_owned()).collect());
 
-        let path = self.get_plist_path();
+        let path = self.get_plist_path()?;
         file.to_writer_xml(std::fs::File::create(&path).wrap_err("Error creating config file")?)
             .wrap_err("Error writing config file")?;
 
@@ -60,9 +82,12 @@ impl ServiceManager for Manager {
     }
 
     fn uninstall(&self) -> Result<()> {
-        let path = self.get_plist_path();
+        let path = self.get_plist_path()?;
         self.run_launchctl(vec!["unload", &path.to_string_lossy()])?;
-        fs::remove_file(path).wrap_err("Error removing config file")?;
+        if path.exists() {
+            fs::remove_file(path).wrap_err("Error removing config file")?;
+        }
+
         Ok(())
     }
 
@@ -78,7 +103,7 @@ impl ServiceManager for Manager {
 
     fn query_status(&self) -> Result<ServiceStatus> {
         let output = self.run_launchctl(vec!["print", &format!("system/{}", self.config.name)])?;
-        if output.starts_with("Could not find service") {
+        if output.starts_with("bad request.") {
             return Ok(ServiceStatus::NotInstalled);
         }
         let s = output
@@ -93,7 +118,9 @@ impl ServiceManager for Manager {
                     .to_string()
             })
             .collect::<Vec<_>>();
-
+        if s.len() == 0 {
+            return Ok(ServiceStatus::Stopped);
+        }
         if s[0].trim() == "running" {
             Ok(ServiceStatus::Started)
         } else {
